+++ src/Resources/public/js/backend.js
@@ -1,0 +1,18 @@
+//(function($) {
+//	$(document).ready(function() {
+//	    $(".markiertIcon img").on('click', function(event) {
+//	        alert("Submit button is clicked!");
+//	        //event.preventDefault();
+//	    });
+//	});
+//})(jQuery); 
+//
+
+function markiertSetzen(feld, value)
+{
+	alert('Hallo');
+	console.log('Hallo');
+	$(document).ready(function(){
+	    $('p').css('color', 'orange');
+	});
+}